+++ CHANGELOG.md
@@ -2,9 +2,16 @@
 
 ## 0.0.19 (Unreleased)
 
+### Bugs Fixed
+
+- Fixes Service Bus host name parameter description. https://github.com/Azure/azure-mcp/pull/209/
+- Updates the usage patterns of Azure Developer CLI (azd) when invoked from MCP. https://github.com/Azure/azure-mcp/pull/203
+
 ### Features Added
 
 - Support for Azure Redis Caches and Clusters https://github.com/Azure/azure-mcp/pull/198
+
+### Other Changes
 
 ## 0.0.18 (2025-05-14)
 
@@ -17,6 +24,7 @@
 ### Bugs Fixed
 
 - Added an opt-in timeout for browser-based authentication to handle cases where the process waits indefinitely if the user closes the browser. https://github.com/Azure/azure-mcp/pull/189
+
 
 ## 0.0.16 (2025-05-13)